Remove logic from top actions
