+++ pages/_app.tsx
@@ -9,7 +9,7 @@ import {
     SolletWalletAdapter,
     TorusWalletAdapter,
 } from '@solana/wallet-adapter-wallets'
-import { Navbar, TopActions, Footer, CheckConnectedWallet } from 'components'
+import { Navbar, Footer, CheckConnectedWallet } from 'components'
 import type { AppProps } from 'next/app'
 import { useMemo } from 'react'
 import { RecoilRoot } from 'recoil'
@@ -41,7 +41,6 @@ function MyApp({ Component, pageProps }: AppProps) {
                         <Navbar />
                         <CheckConnectedWallet>
                             <div className='container-xl p-responsive height-full my-11'>
-                                <TopActions />
                                 <Component {...pageProps} />
                             </div>
                         </CheckConnectedWallet>
