+++ pages/candy-machines/[candy_machine_ID].tsx
@@ -1,11 +1,11 @@
 import { AnchorProvider, Program } from '@project-serum/anchor'
-import { Button, Spinner, Text } from '@primer/react'
+import { Button, Link, Spinner, Text } from '@primer/react'
 import { CANDY_MACHINE_PROGRAM_V2_ID } from 'lib/candy-machine/constants'
 import { getAllNftsByCM, getNftByMint } from 'lib/nft/actions'
 import { IFetchedCandyMachineConfig } from 'lib/candy-machine/interfaces'
 import { Nft } from 'lib/nft/interfaces'
 import { nftsState } from 'lib/recoil-store/atoms'
-import { NftCard, RefreshButton, UpdateCandyMachine } from 'components'
+import { NftCard, Popup, RefreshButton, TopActions, UpdateCandyMachine } from 'components'
 import { useAnchorWallet } from '@solana/wallet-adapter-react'
 import { useEffect, useState } from 'react'
 import { useMintCandyMachine, useRPC } from 'hooks'
@@ -14,6 +14,9 @@ import { useRouter } from 'next/router'
 import { PublicKey } from '@solana/web3.js'
 import Head from 'next/head'
 import type { NextPage } from 'next'
+import DeleteCandyMachine from 'components/DeleteCandyMachine'
+import VerifyCandyMachine from 'components/VerifyCandyMachine'
+import { LinkExternalIcon } from '@primer/octicons-react'
 
 const CandyMachine: NextPage = () => {
     const router = useRouter()
@@ -21,11 +24,10 @@ const CandyMachine: NextPage = () => {
     const candyMachineAccount = router.query.candy_machine_ID as string
 
     const anchorWallet = useAnchorWallet()
-    const { connection } = useRPC()
+    const { connection, isDevnet } = useRPC()
     const [candyMachineConfig, setCandyMachineConfig] = useState<IFetchedCandyMachineConfig>()
     const [error, setError] = useState('')
     const [isLoadingNfts, setIsLoadingNfts] = useState(false)
-    const [initialLoad, setInitialLoad] = useState(false)
     const [nfts, setNfts] = useState<Nft[]>([])
     const [mintedNfts, setMintedNfts] = useState<Nft[]>([])
     const [collectionNft, setCollectionNft] = useState<Nft>()
@@ -35,6 +37,8 @@ const CandyMachine: NextPage = () => {
     const [hasCollection, setHasCollection] = useState<boolean>(false)
     const [nftsRecoilState, setNftsRecoilState] = useRecoilState(nftsState)
     const [cache, setCache] = useState<File>()
+    const [isDeleteOpen, setIsDeleteOpen] = useState(false)
+    const [isVerifyOpen, setIsVerifyOpen] = useState(false)
 
     const viewNfts = async (e: React.ChangeEvent<HTMLInputElement>) => {
         setIsLoadingNfts(true)
@@ -148,6 +152,23 @@ const CandyMachine: NextPage = () => {
                 <meta name='description' content='Generated by create next app' />
                 <link rel='icon' href='/favicon.ico' />
             </Head>
+            <TopActions>
+                <Button variant='danger' onClick={() => setIsDeleteOpen(true)}>
+                    Delete
+                </Button>
+                <Button variant='outline' onClick={() => setIsVerifyOpen(true)}>
+                    Verify
+                </Button>
+                <Button leadingIcon={LinkExternalIcon}>
+                    <Link
+                        target='_blank'
+                        href={`https://solscan.io/account/${candyMachineAccount}?${isDevnet}`}
+                        sx={{ textDecoration: 'none', color: '#24292F' }}
+                    >
+                        View in Solscan
+                    </Link>
+                </Button>
+            </TopActions>
             <div className='d-flex flex-justify-center flex-column'>
                 <h2 className='my-5 wb-break-all'>[CM]: {candyMachineAccount}</h2>
                 {isLoading && loadingText}
@@ -309,6 +330,17 @@ const CandyMachine: NextPage = () => {
                             )}
                         </div>
                     </>
+                )}
+
+                {isDeleteOpen && (
+                    <Popup onClose={() => setIsDeleteOpen(false)} title='Delete Candy Machine' size='small'>
+                        <DeleteCandyMachine candyMachineAccount={candyMachineAccount as string} />
+                    </Popup>
+                )}
+                {isVerifyOpen && (
+                    <Popup onClose={() => setIsVerifyOpen(false)} title={`Verify Candy Machine`} size='small'>
+                        <VerifyCandyMachine candyMachineAccount={candyMachineAccount as string} />
+                    </Popup>
                 )}
             </div>
         </>
