+++ pages/candy-machines/index.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import { useWallet } from '@solana/wallet-adapter-react'
-import { CandyMachineList, CreateCandyMachine, Popup } from 'components'
+import { CandyMachineList, CreateCandyMachine, Popup, SearchBar, RefreshButton, TopActions } from 'components'
 import { useRPC } from 'hooks'
 import { candyMachinesState, candyMachineSearchState } from 'lib/recoil-store/atoms'
 import { fetchCandyMachineAccounts } from 'lib/utils'
@@ -9,16 +9,19 @@ import Head from 'next/head'
 import { useEffect, useState } from 'react'
 import { useRecoilState } from 'recoil'
 import { Button, Spinner } from '@primer/react'
-import { useRecoilValue } from 'recoil'
+import { useRecoilValue, useSetRecoilState } from 'recoil'
 
 const ManageCandyMachines: NextPage = () => {
     const [accounts, setAccounts] = useRecoilState(candyMachinesState)
     const [isLoading, setIsLoading] = useState(false)
+    const [isRefreshLoading, setIsRefreshLoading] = useState(false)
     const [error, setError] = useState(false)
     const { connection } = useRPC()
     const { publicKey } = useWallet()
-    const [isOpen, setIsOpen] = useState(false)
+    const [isCreateOpen, setIsCreateOpen] = useState(false)
+    const [searchValue, setSearchValue] = useRecoilState(candyMachineSearchState)
     const searchInput = useRecoilValue(candyMachineSearchState)
+    const setCandyMachines = useSetRecoilState(candyMachinesState)
 
     const searchResults = accounts.filter((account) => {
         return account.toLowerCase().includes(searchInput.trim().toLowerCase())
@@ -32,6 +35,18 @@ const ManageCandyMachines: NextPage = () => {
         } catch (err) {
             console.error(err)
         }
+    }
+
+    const refreshCandyMachines = async () => {
+        setIsRefreshLoading(true)
+        if (!connection) return
+        try {
+            const candyMachines = await fetchCandyMachineAccounts(connection, publicKey!)
+            setCandyMachines(candyMachines)
+        } catch (e) {
+            console.error(e)
+        }
+        setIsRefreshLoading(false)
     }
 
     useEffect(() => {
@@ -62,6 +77,17 @@ const ManageCandyMachines: NextPage = () => {
                 <meta name='description' content='Generated by create next app' />
                 <link rel='icon' href='/favicon.ico' />
             </Head>
+            <TopActions>
+                <SearchBar
+                    searchValue={searchValue}
+                    setSearchValue={setSearchValue}
+                    placeholderText='Search candy machine'
+                />
+                <Button variant='primary' onClick={() => setIsCreateOpen(true)} sx={{ textTransform: 'capitalize' }}>
+                    Create candy machine
+                </Button>
+                <RefreshButton onClick={refreshCandyMachines} isLoading={isRefreshLoading} />
+            </TopActions>
             <div className='d-flex flex-column'>
                 <div className='d-flex my-5 flex-items-center'>
                     <h2>Candy Machines ·</h2>
@@ -80,19 +106,19 @@ const ManageCandyMachines: NextPage = () => {
                         <Button
                             className='color-bg-transparent text-semibold'
                             variant='invisible'
-                            onClick={() => setIsOpen(true)}
+                            onClick={() => setIsCreateOpen(true)}
                             sx={{ textTransform: 'capitalize', padding: '0' }}
                         >
                             Create one?
                         </Button>
-                        {isOpen && (
-                            <Popup onClose={() => setIsOpen(false)} title='Create Candy Machine' size='large'>
-                                <CreateCandyMachine />
-                            </Popup>
-                        )}
                     </div>
                 ) : (
                     <CandyMachineList candyMachineAccounts={searchResults} />
+                )}
+                {isCreateOpen && (
+                    <Popup onClose={() => setIsCreateOpen(false)} title='Create Candy Machine' size='large'>
+                        <CreateCandyMachine />
+                    </Popup>
                 )}
             </div>
         </>
